refactor(split): move central split run out of Bluetooth directory
