+++ app/include/zmk/split/service.h
@@ -31,6 +31,11 @@ struct zmk_split_run_behavior_payload {
     char behavior_dev[ZMK_SPLIT_RUN_BEHAVIOR_DEV_LEN];
 } __packed;
 
+struct zmk_split_run_behavior_payload_wrapper {
+    uint8_t source;
+    struct zmk_split_run_behavior_payload payload;
+};
+
 int zmk_split_position_pressed(uint8_t position);
 int zmk_split_position_released(uint8_t position);
 int zmk_split_sensor_triggered(uint8_t sensor_index,
